abseil-cpp: build with c++17 by default
Abseil provides compatibility types for c++17. These types are defined
differently depending on the value of `-std=c++XX`. When these types are
used in public interfaces they can cause compilation or linker failures
when a dependent library is compiled with a different standard than
abseil itself.

The workaround before this commit was to create multiple configurations
of abseil depending on the selected c++ standard of the downstream using
package. This approach has two drawbacks:
* It can easily create situations in which multiple configurations of
  abseil exist in a single build closure, and that requires rather
  creative workarounds.
* The overridden setting for -std=c++XX must be propagated downwards
  along the depency chain.

This Commit does not fully solve either problem, but it makes it easier
to deal with. By defaulting to c++17 and aligning the package tree below
abseil to that we can drop complicated logic that inspects intermediate
derivations in the sub-tree that override `cxxStandard`, and prevent
most cases of disagreement.